Move app reducer ownership into app delegate to simplify initialization.
