+++ Keyb/KeybApp.swift
@@ -13,10 +13,18 @@ struct KeybApp: App {
 
     @NSApplicationDelegateAdaptor(AppDelegate.self) var appDelegate
 
+    var body: some Scene {
+        WindowGroup {
+            UserInterfaceView(store: appDelegate.store)
+        }
+    }
+}
+
+final class AppDelegate: NSObject, NSApplicationDelegate {
+
     let store = Store(
         initialState: UserInterfaceState(mode: .noAccessibilityPermission(.hasNotPromptedYet)),
-        reducer: userInterfaceReducer
-            .combined(with: appDelegateReducer.pullback(state: \.appDelegate, action: /.self, environment: { _ in })),
+        reducer: appReducer,
         environment: .init(
             accessibilityClient: .live,
             eventHandlerClient: .live(with: eventHandler),
@@ -24,45 +32,10 @@ struct KeybApp: App {
         )
     )
 
-    init() {
-        appDelegate.store = store.scope(state: \.appDelegate)
-    }
-
-    var body: some Scene {
-        WindowGroup {
-            UserInterfaceView(store: store)
-        }
-    }
-}
-
-extension UserInterfaceState {
-    var appDelegate: AppDelegate.State {
-        get {
-            .init(
-                isDockMenuItemChecked: isRunning,
-                hasPermission: mode.hasPermission
-            )
-        }
-        set {
-            isRunning = newValue.isDockMenuItemChecked
-        }
-    }
-}
-
-final class AppDelegate: NSObject, NSApplicationDelegate {
-
-    struct State: Equatable {
-        var isDockMenuItemChecked: Bool
-        var hasPermission: Bool
-    }
-
-    var store: Store<State, UserInterfaceAction>?
+    lazy var viewStore = ViewStore(store.scope(state: \.appDelegate))
 
     func applicationDockMenu(_ sender: NSApplication) -> NSMenu? {
-        guard
-            let store = store,
-            case let viewStore = ViewStore(store),
-            viewStore.hasPermission
+        guard viewStore.hasPermission
         else { return nil }
 
         let isChecked = viewStore.isDockMenuItemChecked
@@ -80,21 +53,11 @@ final class AppDelegate: NSObject, NSApplicationDelegate {
     }
 
     @objc func enable() {
-        store.map(ViewStore.init)?.send(.changeObservingState(observing: true))
+        viewStore.send(.changeObservingState(observing: true))
 
     }
 
     @objc func disable() {
-        store.map(ViewStore.init)?.send(.changeObservingState(observing: false))
-    }
-}
-
-let appDelegateReducer = Reducer<AppDelegate.State, UserInterfaceAction, Void> { state, action, _ in
-    switch action {
-    case .changeObservingState(observing: let observing):
-        state.isDockMenuItemChecked = observing
-        return .none
-    default:
-        return .none
+        viewStore.send(.changeObservingState(observing: false))
     }
 }
